+++ src/components/composite/SidePanel/SoundsPanel.tsx
@@ -15,6 +15,7 @@ import { AddSetDialog } from "./AddSetDialog";
 import { AddToSetDialog } from "./AddToSetDialog";
 import { AddTagsDialog } from "./AddTagsDialog";
 import { DownloadDialog } from "@/components/modals/DownloadDialog";
+import { DownloadStatusButton } from "@/components/composite/DownloadManager/DownloadStatusButton";
 import { FolderBrowser } from "./FolderBrowser";
 import { SoundList } from "./SoundList";
 import { ConfirmRemoveMissingDialog } from "@/components/modals/ConfirmRemoveMissingDialog";
@@ -114,6 +115,7 @@ export function SoundsPanel() {
           </TooltipTrigger>
           <TooltipContent>Download from URL</TooltipContent>
         </Tooltip>
+        <DownloadStatusButton />
         <div className="relative ml-auto flex items-center">
           <HugeiconsIcon icon={Search01Icon} size={14} className="absolute left-2.5 text-white/50 pointer-events-none" />
           <Input